refactor(exec): Move ExecOptions back to index.ts
